rename clusterinfrarefstatus resource to statuscondition
