+++ service/controller/resource/statuscondition/resource.go
@@ -1,4 +1,4 @@
-package clusterinfrarefstatus
+package statuscondition
 
 import (
 	infrastructurev1alpha2 "github.com/giantswarm/apiextensions/pkg/apis/infrastructure/v1alpha2"
@@ -8,7 +8,7 @@ import (
 )
 
 const (
-	Name = "clusterinfrarefstatus"
+	Name = "statuscondition"
 )
 
 type Config struct {
